requirements.txt
aiohttp==3.12.15
feedparser==6.0.11
html2text==2025.4.15
httpx==0.28.1
nonebot==1.9.1
Pillow==11.3.0
pydantic==2.11.7
Requests==2.32.5